\name{logLinear}
\alias{logLinear}
\title{log-linear terms}
\usage{
  logLinear(data, f, fixed = NULL, starting = NULL,
    prefixes = "beta")
}
\arguments{
  \item{data}{data frame for evaluation}

  \item{f}{formula}

  \item{fixed}{list of fixed parameters}

  \item{starting}{list of starting values}

  \item{prefixes}{prefix of the variable name}
}
\description{
  log-linear terms
}

